% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{assert}
\alias{assert}
\title{assert}
\usage{
assert(expr, error)
}
\arguments{
\item{expr}{expression to check}

\item{error}{error message to display}
}
\description{
Verify a condition and stop the execution if it is not met
}
\examples{

\dontrun{
assert(1 == 2, "1 is not 2!!")
}
}
