% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dhms}
\alias{dhms}
\title{dhms}
\usage{
dhms(t)
}
\arguments{
\item{t}{Time in seconds}
}
\value{

}
\description{
Convert a time expressed in seconds into hour:minute:second format
}
\examples{
dhms(86400)
}
