% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_datetime_from_file}
\alias{get_datetime_from_file}
\title{get_datetime_from_file}
\usage{
get_datetime_from_file(filename)
}
\arguments{
\item{filename}{file name to process}
}
\description{
On Sansa devices with rockbox firmware, files are saved in a format where the recording start
date and time can be retrieved
}
\examples{

get_datetime_from_file("R_MIC_091020-123456.mp3")

}
