% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{hms}
\alias{hms}
\title{hms}
\usage{
hms(t)
}
\arguments{
\item{t}{Time in seconds}
}
\value{
a string of time formated
}
\description{
Convert a time expressed in seconds into hour:minute:second format
}
\examples{
hms(86400)
}
