% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lutl.R
\name{load_lutl}
\alias{load_lutl}
\title{load_lutl}
\usage{
load_lutl(x)
}
\arguments{
\item{x}{`lutl` object filename, like `*.Rds`}
}
\description{
Loads a previously saved 'lutl' object
}
\examples{
\dontrun{
load_lutl("data.Rds")
}
}
