% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loudness-package.R
\docType{package}
\name{loudness}
\alias{loudness}
\title{loudnes}
\description{
This package contains utility function for some common data reshaping
operations, and also some HTML utility functions for greater control over
R markdown documents.
}
\seealso{
\code{\link{html}}
}
