% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lutl.R
\name{lutl_from_file}
\alias{lutl_from_file}
\title{lutl_from_file}
\usage{
lutl_from_file(
  file,
  skip_first_n = 3,
  start_date = NA,
  from = NA,
  duration = NA,
  group_by_n = 1,
  adjfun = mean,
  compact_fmt = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{file}{Any file with an audio track that is consistent with ffmpeg}

\item{skip_first_n}{Skip the first n observations. On at least one Sansa Fuze device, the first observations
of the recorded audio are basically silent, so they can be ignored.}

\item{start_date}{Date and time of the start of the audio recording. In case you do not set this parameter,
it tries to retrieve the file name information, for now only valid for the devices with
the rockbox firmware.}

\item{from}{From "hour:minute:second"}

\item{duration}{Length in "Hour:minute:second"}

\item{group_by_n}{Each observation corresponds to one tenth of a second, with this parameter we can group
them to reduce the final size of the data. Example `group_by_n = 10` corresponds to one
observation every second.}

\item{adjfun}{Summation or adjustment function when grouping observations By default, `mean` is used}

\item{compact_fmt}{Use compact format? (default = TRUE)}

\item{verbose}{default FALSE}
}
\value{
a lutl object
}
\description{
Loudness Units Time Line from audo file. A lutl object is a timeline of sound levels
with a resolution of tenths of a second. The data is obtained by processing an audio
file with ffmpeg and the EBUR128 plug-in.
}
\details{
EBU R 128 is a recommendation for loudness normalisation and maximum level of audio
signals. It is primarily followed during audio mixing of television and radio
programmes and adopted by broadcasters to measure and control programme loudness.
It was first issued by the European Broadcasting Union in August 2010 and most
recently revised in June 2014. R 128 employs an international standard for measuring
audio loudness, stated in the ITU-R BS.1770 recommendation and using the loudness
measures LU (loudness units) and LUFS (loudness units referenced to full scale),
specifically created with this purpose.[3] The EBU Tech 3341 document further clarified
loudness metering implementation and practices in 2016.
}
\examples{
\dontrun{
data <- lutl_from_file('~/Descargas/ejemplo.wav')
data <- lutl_from_file('~/Descargas/ejemplo.wav', start_date = "2020-10-02 16:45:34")
}
}
