% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{parse_ffmpeg_out}
\alias{parse_ffmpeg_out}
\title{parse_ffmpeg_out}
\usage{
parse_ffmpeg_out(filepath, nlines = 1, compact_fmt = TRUE)
}
\arguments{
\item{filepath}{path to the file to be processed}

\item{nlines}{read blocks of so many lines}

\item{compact_fmt}{Compact format for output (Default TRUE)}
}
\value{
a data.frame with loudness data
}
\description{
Interprets the log file generated by ffmpeg with the loudness detail
}
\examples{

\dontrun{
parse_ffmpeg_out('ffmpeg.out', nlines = 1000)
}
}
