% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{parse_ffmpeg_out_fast}
\alias{parse_ffmpeg_out_fast}
\title{parse_ffmpeg_out_fast}
\usage{
parse_ffmpeg_out_fast(filepath, nlines = 1, compact_fmt = TRUE)
}
\arguments{
\item{filepath}{path to the file to be processed}

\item{nlines}{number of buffered lines}

\item{compact_fmt}{Compact format for output (Default TRUE)}
}
\value{
a data.frame with loudness data
}
\description{
Interprets the log file generated by ffmpeg with the loudness detail
}
\examples{

\dontrun{
parse_ffmpeg_out_fast('ffmpeg.out')
}
}
