% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play.lutl}
\alias{play.lutl}
\title{play.lutl}
\usage{
\method{play}{lutl}(x, from = NA, duration = NA, verbose = FALSE, ...)
}
\arguments{
\item{x}{A `lutl` object}

\item{from}{From "hour:minute:second" (Default `NA`, from begining of timeline)}

\item{duration}{Length in "Hour:minute:second" (Default `NA`, all timeline)}

\item{verbose}{be verbose (default FALSE)}

\item{...}{Anything else}
}
\description{
Play a lutl object
}
\examples{
\dontrun{
data <- lutl_from_file('~/Descargas/ejemplo.wav')
play(data)
}
}
