% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play_secs}
\alias{play_secs}
\title{play_secs}
\usage{
play_secs(x, secs = 1, from = NA, verbose = FALSE, ...)
}
\arguments{
\item{x}{A `lutl` object}

\item{secs}{number of seconds to play}

\item{from}{From "hour:minute:second" (Default `NA`, from begining of timeline)}

\item{verbose}{be verbose (default FALSE)}

\item{...}{Anything else}
}
\description{
play some seconds of a lutl object
}
\examples{
\dontrun{
data <- lutl_from_file('~/Descargas/ejemplo.wav')
play_secs(data, 5)
}
}
