% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_density}
\alias{plot_density}
\title{plot_density}
\usage{
plot_density(x, from = NA, duration = NA, ...)
}
\arguments{
\item{x}{A `lutl` object}

\item{from}{From "hour:minute:second" (Default `NA`, from begining of timeline)}

\item{duration}{Length in "Hour:minute:second" (Default `NA`, all timeline)}

\item{...}{Anything}
}
\description{
A Kernel Density of LUFS timeline
}
\examples{
\dontrun{
plot_density(data)
}
}
