% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeline}
\alias{plot_timeline}
\title{plot_timeline}
\usage{
plot_timeline(x, percentil = NA, nbreaks = 10)
}
\arguments{
\item{x}{a `lutl` object}

\item{percentil}{Limit the plot to sounds above this percentile}
}
\description{
It is a scatter plot of sound levels on the timeline
}
\examples{
\dontrun{
plot_timeline(data, percentil=99)
}
}
