% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/read.R
\name{read}
\alias{read}
\title{read}
\usage{
read(file)

read(file)
}
\arguments{
\item{file}{Path to a file.}

\item{path}{File path}

\item{lines}{As text lines file}
}
\description{
Generic fast read

These functions read a file into memory. We memory map the file for fast I/O.
The file is read in as a character vector (length one for \code{read},
length \code{n} for \code{readlines}).
}
