% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readlines}
\alias{readlines}
\title{Read lines ofa File}
\usage{
readlines(file)
}
\arguments{
\item{file}{Path to a file.}
}
\description{
Read lines ofa File
}
\examples{
p <- file.path( R.home(), "NEWS" )
if (file.exists(p))
  stopifnot( identical( readLines(p), readlines(p) ) )
}
