% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{run_cmd}
\alias{run_cmd}
\title{run_cmd}
\usage{
run_cmd(
  cmd,
  wait = TRUE,
  echo = FALSE,
  ignore.stdout = FALSE,
  ignore.stderr = FALSE
)
}
\arguments{
\item{cmd}{Command to execute}

\item{wait}{Wait command to finish (default TRUE)}

\item{echo}{Print command in console (default FALSE)}

\item{ignore.stdout, ignore.stderr}{a logical (not NA) indicating whether messages written to ‘stdout’
or ‘stderr’ should be ignored.}
}
\value{
a value returned from `system` or `shell`
}
\description{
Execute a command in a way that is consistent with both Windows and Linux platforms
}
\examples{

run_cmd("ls")

}
