% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lutl.R
\name{save_lutl}
\alias{save_lutl}
\title{save_lutl}
\usage{
save_lutl(x, ...)
}
\arguments{
\item{x}{a `lutl` object}

\item{...}{Anything}
}
\description{
Save a `lutl` object in native R format RDS
}
\examples{
\dontrun{
save_lutl(data)
}
}
