% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{try_require}
\alias{try_require}
\title{try_require}
\usage{
try_require(package, fun)
}
\arguments{
\item{package}{name of the package to load}

\item{fun}{name of function that needs the package}
}
\description{
Attempts to load a package in case an error is issued
}
\examples{

try_require("stats", "my_function")

}
