% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_datetime_from_file}
\alias{get_datetime_from_file}
\title{get_datetime_from_file}
\usage{
get_datetime_from_file(filename)
}
\arguments{
\item{filename}{file name to process}
}
\description{
get_datetime_from_file
}
\examples{

get_datetime_from_file("R_MIC_091020-123456.mp3")

}
